requirements.txt
biopython>=1.79
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
requests>=2.25.0
pybedtools>=0.9.0 